refactor: split AssignmentRoutes.swift (2588 lines) into 6 focused files
Pure mechanical split — no behaviour changes.

- AssignmentRoutes.swift        → 1040 lines (boot + core assignment CRUD/lifecycle)
- AssignmentRoutes+Sections.swift  → 175 lines (section create/rename/reorder/delete/move)
- AssignmentRoutes+Submissions.swift → 365 lines (grade CSV, submissions, history, retest)
- AssignmentRoutes+Enrollment.swift →  89 lines (open-enrolment toggle, bulk CSV enroll)
- AssignmentHelpers.swift       → 835 lines (all free utility functions + helper types)
- AssignmentContextTypes.swift  → 134 lines (all Encodable view-context structs)

Two incidental fixes: renamed SubmissionHistoryRow → AssignmentSubmissionHistoryRow
to avoid collision with WebRoutes.swift's private version; removed duplicate
gradePercentFromCollectionJSON from WebRoutes.swift now that it's internal in
AssignmentHelpers.swift.
