+++ Sources/APIServer/Routes/Web/AssignmentRoutes.swift
@@ -275,360 +275,6 @@ struct AssignmentRoutes: RouteCollection {
         return try await req.view.render("assignments", ctx).encodeResponse(for: req)
     }
 
-    // MARK: - GET /assignments/grades.csv
-
-    @Sendable
-    func exportGradesCSV(req: Request) async throws -> Response {
-        let user = try req.auth.require(APIUser.self)
-        let courseState = try await req.resolveActiveCourse(for: user)
-
-        // Only include students enrolled in the active course (if one is set).
-        let students: [APIUser]
-        if let activeCourseUUID = courseState.activeCourseUUID {
-            let enrolledUserIDs = try await APICourseEnrollment.query(on: req.db)
-                .filter(\.$course.$id == activeCourseUUID)
-                .all()
-                .map { $0.userID }
-            let enrolledSet = Set(enrolledUserIDs)
-            let allStudents = try await APIUser.query(on: req.db)
-                .filter(\.$role == "student")
-                .sort(\.$username, .ascending)
-                .all()
-            students = allStudents.filter { u in u.id.map { enrolledSet.contains($0) } ?? false }
-        } else {
-            students = try await APIUser.query(on: req.db)
-                .filter(\.$role == "student")
-                .sort(\.$username, .ascending)
-                .all()
-        }
-
-        let assignments: [APIAssignment]
-        if let activeCourseUUID = courseState.activeCourseUUID {
-            assignments = try await APIAssignment.query(on: req.db)
-                .filter(\.$courseID == activeCourseUUID)
-                .all()
-        } else {
-            assignments = try await APIAssignment.query(on: req.db).all()
-        }
-        let setupIDs = Set(assignments.map(\.testSetupID))
-        let setups = setupIDs.isEmpty
-            ? []
-            : try await APITestSetup.query(on: req.db)
-                .filter(\.$id ~~ setupIDs)
-                .all()
-        let setupByID = Dictionary(uniqueKeysWithValues: setups.compactMap { setup in
-            setup.id.map { ($0, setup) }
-        })
-
-        let sortedAssignments = assignments.sorted { lhs, rhs in
-            switch (lhs.sortOrder, rhs.sortOrder) {
-            case let (l?, r?) where l != r:
-                return l < r
-            default:
-                let lhsCreated = setupByID[lhs.testSetupID]?.createdAt ?? .distantPast
-                let rhsCreated = setupByID[rhs.testSetupID]?.createdAt ?? .distantPast
-                if lhsCreated != rhsCreated { return lhsCreated > rhsCreated }
-                return lhs.testSetupID < rhs.testSetupID
-            }
-        }
-
-        let studentIDs = Set(students.compactMap(\.id))
-        let submissionRows = (setupIDs.isEmpty || studentIDs.isEmpty)
-            ? []
-            : try await APISubmission.query(on: req.db)
-                .filter(\.$kind == APISubmission.Kind.student)
-                .filter(\.$testSetupID ~~ setupIDs)
-                .filter(\.$userID ~~ studentIDs)
-                .all()
-        let submissions = submissionRows.compactMap { row -> (id: String, userID: UUID, setupID: String)? in
-            guard let id = row.id, let userID = row.userID else { return nil }
-            return (id, userID, row.testSetupID)
-        }
-        let submissionIDs = submissions.map(\.id)
-
-        let results = submissionIDs.isEmpty
-            ? []
-            : try await APIResult.query(on: req.db)
-                .filter(\.$submissionID ~~ submissionIDs)
-                .sort(\.$receivedAt, .descending)
-                .all()
-        var preferredResultBySubmissionID: [String: APIResult] = [:]
-        for result in results {
-            let key = result.submissionID
-            if let existing = preferredResultBySubmissionID[key] {
-                let existingSource = existing.source ?? "worker"
-                let candidateSource = result.source ?? "worker"
-                if existingSource == "worker" { continue }
-                if candidateSource == "worker" {
-                    preferredResultBySubmissionID[key] = result
-                }
-            } else {
-                preferredResultBySubmissionID[key] = result
-            }
-        }
-
-        var bestGradeByUserAndSetup: [String: Int] = [:]
-        for submission in submissions {
-            guard let result = preferredResultBySubmissionID[submission.id],
-                  let grade = gradePercentFromCollectionJSON(result.collectionJSON) else {
-                continue
-            }
-            let key = "\(submission.userID.uuidString.lowercased())::\(submission.setupID)"
-            let prior = bestGradeByUserAndSetup[key] ?? -1
-            if grade > prior {
-                bestGradeByUserAndSetup[key] = grade
-            }
-        }
-
-        var lines: [String] = []
-        let header = ["student_id"] + sortedAssignments.map { $0.title }
-        lines.append(header.map(csvEscaped).joined(separator: ","))
-
-        for student in students {
-            guard let userID = student.id else { continue }
-            var row: [String] = [student.username]
-            for assignment in sortedAssignments {
-                let key = "\(userID.uuidString.lowercased())::\(assignment.testSetupID)"
-                if let grade = bestGradeByUserAndSetup[key] {
-                    row.append("\(grade)%")
-                } else {
-                    row.append("")
-                }
-            }
-            lines.append(row.map(csvEscaped).joined(separator: ","))
-        }
-
-        let csv = lines.joined(separator: "\n") + "\n"
-        let timestamp = Int(Date().timeIntervalSince1970)
-        let response = Response(status: .ok)
-        response.headers.replaceOrAdd(name: .contentType, value: "text/csv; charset=utf-8")
-        response.headers.replaceOrAdd(
-            name: .contentDisposition,
-            value: "attachment; filename=\"grades-\(timestamp).csv\""
-        )
-        response.body = .init(string: csv)
-        return response
-    }
-
-    // MARK: - GET /assignments/:assignmentID/submissions
-
-    @Sendable
-    func assignmentSubmissionsPage(req: Request) async throws -> View {
-        let assignmentIDRaw = try assignmentPublicIDParameter(from: req)
-        guard let assignment = try await assignmentByPublicID(assignmentIDRaw, on: req.db) else {
-            throw Abort(.notFound)
-        }
-
-        let students = try await APIUser.query(on: req.db)
-            .filter(\.$role == "student")
-            .sort(\.$username, .ascending)
-            .all()
-        let studentIDs = Set(students.compactMap(\.id))
-
-        let submissions = (studentIDs.isEmpty)
-            ? []
-            : try await APISubmission.query(on: req.db)
-                .filter(\.$testSetupID == assignment.testSetupID)
-                .filter(\.$kind == APISubmission.Kind.student)
-                .filter(\.$userID ~~ studentIDs)
-                .sort(\.$submittedAt, .descending)
-                .all()
-
-        var submissionsByStudentID: [UUID: [APISubmission]] = [:]
-        for row in submissions {
-            guard let userID = row.userID else { continue }
-            submissionsByStudentID[userID, default: []].append(row)
-        }
-
-        let submissionIDs = submissions.compactMap(\.id)
-        let results = submissionIDs.isEmpty
-            ? []
-            : try await APIResult.query(on: req.db)
-                .filter(\.$submissionID ~~ submissionIDs)
-                .sort(\.$receivedAt, .descending)
-                .all()
-        var preferredResultBySubmissionID: [String: APIResult] = [:]
-        for result in results {
-            let key = result.submissionID
-            if let existing = preferredResultBySubmissionID[key] {
-                let existingSource = existing.source ?? "worker"
-                let candidateSource = result.source ?? "worker"
-                if existingSource == "worker" { continue }
-                if candidateSource == "worker" {
-                    preferredResultBySubmissionID[key] = result
-                }
-            } else {
-                preferredResultBySubmissionID[key] = result
-            }
-        }
-
-        let fmt = DateFormatter()
-        fmt.dateStyle = .medium
-        fmt.timeStyle = .short
-
-        let rows = students.compactMap { student -> AssignmentStudentRow? in
-            guard let studentID = student.id else { return nil }
-            let history = submissionsByStudentID[studentID] ?? []
-            let latest = history.first
-            let grade: String = {
-                var best = -1
-                for submission in history {
-                    guard let subID = submission.id,
-                          let result = preferredResultBySubmissionID[subID],
-                          let pct = gradePercentFromCollectionJSON(result.collectionJSON) else {
-                        continue
-                    }
-                    if pct > best { best = pct }
-                }
-                return best >= 0 ? "\(best)%" : "—"
-            }()
-            let inferredName = inferNameFromStudentID(student.username)
-            return AssignmentStudentRow(
-                studentID: student.username,
-                surname: inferredName.surname,
-                givenNames: inferredName.givenNames,
-                gradeText: grade,
-                submissionCount: history.count,
-                hasLatestSubmission: latest != nil,
-                latestSubmissionID: latest?.id ?? "",
-                latestSubmittedAtText: latest?.submittedAt.map { fmt.string(from: $0) } ?? "—",
-                additionalSubmissionCount: max(history.count - 1, 0),
-                fullHistoryURL: "/assignments/\(assignmentIDRaw)/students/\(studentID.uuidString)/history"
-            )
-        }
-
-        return try await req.view.render(
-            "assignment-submissions",
-            AssignmentSubmissionsContext(
-                currentUser: req.currentUserContext,
-                assignmentID: assignmentIDRaw,
-                assignmentTitle: assignment.title,
-                rows: rows
-            )
-        )
-    }
-
-    // MARK: - GET /assignments/:assignmentID/students/:studentID/history
-
-    @Sendable
-    func studentSubmissionHistoryPage(req: Request) async throws -> View {
-        let assignmentIDRaw = try assignmentPublicIDParameter(from: req)
-        guard
-            let assignment = try await assignmentByPublicID(assignmentIDRaw, on: req.db),
-            let studentIDRaw = req.parameters.get("studentID"),
-            let studentID = UUID(uuidString: studentIDRaw),
-            let student = try await APIUser.find(studentID, on: req.db),
-            student.role == "student"
-        else {
-            throw Abort(.notFound)
-        }
-
-        let submissions = try await APISubmission.query(on: req.db)
-            .filter(\.$testSetupID == assignment.testSetupID)
-            .filter(\.$userID == studentID)
-            .filter(\.$kind == APISubmission.Kind.student)
-            .sort(\.$submittedAt, .descending)
-            .all()
-        let submissionIDs = submissions.compactMap(\.id)
-        let results = submissionIDs.isEmpty
-            ? []
-            : try await APIResult.query(on: req.db)
-                .filter(\.$submissionID ~~ submissionIDs)
-                .sort(\.$receivedAt, .descending)
-                .all()
-
-        var preferredResultBySubmissionID: [String: APIResult] = [:]
-        for result in results {
-            let key = result.submissionID
-            if let existing = preferredResultBySubmissionID[key] {
-                let existingSource = existing.source ?? "worker"
-                let candidateSource = result.source ?? "worker"
-                if existingSource == "worker" { continue }
-                if candidateSource == "worker" {
-                    preferredResultBySubmissionID[key] = result
-                }
-            } else {
-                preferredResultBySubmissionID[key] = result
-            }
-        }
-
-        let fmt = DateFormatter()
-        fmt.dateStyle = .medium
-        fmt.timeStyle = .short
-
-        let rows = submissions.map { submission -> SubmissionHistoryRow in
-            let subID = submission.id ?? ""
-            let gradeText: String
-            if let result = preferredResultBySubmissionID[subID],
-               let pct = gradePercentFromCollectionJSON(result.collectionJSON) {
-                gradeText = "\(pct)%"
-            } else {
-                gradeText = "—"
-            }
-            return SubmissionHistoryRow(
-                submissionID: subID,
-                attemptNumber: submission.attemptNumber ?? 1,
-                status: submission.status,
-                submittedAt: submission.submittedAt.map { fmt.string(from: $0) } ?? "—",
-                gradeText: gradeText
-            )
-        }
-
-        return try await req.view.render(
-            "assignment-student-history",
-            AssignmentStudentHistoryContext(
-                currentUser: req.currentUserContext,
-                assignmentID: assignmentIDRaw,
-                assignmentTitle: assignment.title,
-                studentID: student.username,
-                historyPath: "/assignments/\(assignmentIDRaw)/students/\(studentIDRaw)/history",
-                rows: rows
-            )
-        )
-    }
-
-    // MARK: - POST /assignments/:assignmentID/submissions/:submissionID/retest
-
-    @Sendable
-    func retestSubmission(req: Request) async throws -> Response {
-        struct RetestBody: Content {
-            var returnTo: String?
-        }
-
-        let assignmentIDRaw = try assignmentPublicIDParameter(from: req)
-        guard
-            let assignment = try await assignmentByPublicID(assignmentIDRaw, on: req.db),
-            let submissionID = req.parameters.get("submissionID"),
-            let submission = try await APISubmission.find(submissionID, on: req.db)
-        else {
-            throw Abort(.notFound)
-        }
-
-        guard submission.testSetupID == assignment.testSetupID else {
-            throw Abort(.notFound)
-        }
-        guard submission.kind == APISubmission.Kind.student else {
-            throw Abort(.badRequest, reason: "Only student submissions can be re-tested.")
-        }
-
-        // Prevent duplicate queue entries for in-flight jobs.
-        if submission.status != "pending" && submission.status != "assigned" {
-            submission.status = "pending"
-            submission.workerID = nil
-            submission.assignedAt = nil
-            try await submission.save(on: req.db)
-        }
-
-        let body = try? req.content.decode(RetestBody.self)
-        let fallbackPath = "/assignments/\(assignmentIDRaw)/submissions"
-        let redirectPath = sanitizedAssignmentReturnPath(
-            body?.returnTo,
-            assignmentIDRaw: assignmentIDRaw,
-            fallbackPath: fallbackPath
-        )
-        return req.redirect(to: redirectPath)
-    }
-
     // MARK: - GET /assignments/new
 
     @Sendable
@@ -982,171 +628,6 @@ struct AssignmentRoutes: RouteCollection {
             assignment.sortOrder = index + 1
             try await assignment.save(on: req.db)
         }
-        return .ok
-    }
-
-    // MARK: - POST /assignments/sections
-
-    @Sendable
-    func createSection(req: Request) async throws -> Response {
-        struct CreateSectionBody: Content {
-            var name: String
-            var defaultGradingMode: String
-        }
-        let user = try req.auth.require(APIUser.self)
-        let courseState = try await req.resolveActiveCourse(for: user)
-        guard let courseID = courseState.activeCourseUUID else {
-            throw Abort(.badRequest, reason: "No active course selected.")
-        }
-        let body = try req.content.decode(CreateSectionBody.self)
-        let name = body.name.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard !name.isEmpty else {
-            throw Abort(.badRequest, reason: "Section name must not be empty.")
-        }
-        let mode = body.defaultGradingMode
-        guard mode == "browser" || mode == "worker" else {
-            throw Abort(.badRequest, reason: "defaultGradingMode must be 'browser' or 'worker'.")
-        }
-        let maxOrder = try await APICourseSection.query(on: req.db)
-            .filter(\.$courseID == courseID)
-            .max(\.$sortOrder) ?? 0
-        let section = APICourseSection(name: name, defaultGradingMode: mode, sortOrder: maxOrder + 1, courseID: courseID)
-        try await section.save(on: req.db)
-        return req.redirect(to: "/assignments")
-    }
-
-    // MARK: - POST /assignments/sections/reorder
-
-    @Sendable
-    func reorderSections(req: Request) async throws -> HTTPStatus {
-        struct ReorderBody: Content {
-            var sectionIDs: [String]
-        }
-        let user = try req.auth.require(APIUser.self)
-        let courseState = try await req.resolveActiveCourse(for: user)
-        guard let courseID = courseState.activeCourseUUID else { return .ok }
-        let body = try req.content.decode(ReorderBody.self)
-        let uuids = body.sectionIDs.compactMap { UUID(uuidString: $0) }
-        guard uuids.count == body.sectionIDs.count, !uuids.isEmpty else {
-            throw Abort(.badRequest, reason: "Invalid section ID in reorder payload.")
-        }
-        let sections = try await APICourseSection.query(on: req.db)
-            .filter(\.$courseID == courseID)
-            .filter(\.$id ~~ uuids)
-            .all()
-        guard sections.count == uuids.count else {
-            throw Abort(.badRequest, reason: "Section set mismatch in reorder payload.")
-        }
-        let byID = Dictionary(uniqueKeysWithValues: sections.compactMap { s -> (UUID, APICourseSection)? in
-            guard let id = s.id else { return nil }
-            return (id, s)
-        })
-        for (index, uuid) in uuids.enumerated() {
-            guard let section = byID[uuid] else { continue }
-            section.sortOrder = index + 1
-            try await section.save(on: req.db)
-        }
-        return .ok
-    }
-
-    // MARK: - POST /assignments/sections/:sectionID/rename
-
-    @Sendable
-    func renameSection(req: Request) async throws -> Response {
-        struct RenameSectionBody: Content {
-            var name: String
-            var defaultGradingMode: String
-        }
-        let user = try req.auth.require(APIUser.self)
-        let courseState = try await req.resolveActiveCourse(for: user)
-        guard let courseID = courseState.activeCourseUUID else {
-            throw Abort(.badRequest, reason: "No active course selected.")
-        }
-        guard let sectionIDStr = req.parameters.get("sectionID"),
-              let sectionUUID = UUID(uuidString: sectionIDStr) else {
-            throw Abort(.notFound)
-        }
-        guard let section = try await APICourseSection.find(sectionUUID, on: req.db),
-              section.courseID == courseID else {
-            throw Abort(.notFound)
-        }
-        let body = try req.content.decode(RenameSectionBody.self)
-        let name = body.name.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard !name.isEmpty else {
-            throw Abort(.badRequest, reason: "Section name must not be empty.")
-        }
-        let mode = body.defaultGradingMode
-        guard mode == "browser" || mode == "worker" else {
-            throw Abort(.badRequest, reason: "defaultGradingMode must be 'browser' or 'worker'.")
-        }
-        section.name = name
-        section.defaultGradingMode = mode
-        try await section.save(on: req.db)
-        return req.redirect(to: "/assignments")
-    }
-
-    // MARK: - POST /assignments/sections/:sectionID/delete
-
-    @Sendable
-    func deleteSection(req: Request) async throws -> Response {
-        let user = try req.auth.require(APIUser.self)
-        let courseState = try await req.resolveActiveCourse(for: user)
-        guard let courseID = courseState.activeCourseUUID else {
-            throw Abort(.badRequest, reason: "No active course selected.")
-        }
-        guard let sectionIDStr = req.parameters.get("sectionID"),
-              let sectionUUID = UUID(uuidString: sectionIDStr) else {
-            throw Abort(.notFound)
-        }
-        guard let section = try await APICourseSection.find(sectionUUID, on: req.db),
-              section.courseID == courseID else {
-            throw Abort(.notFound)
-        }
-        // FK SET NULL: assignments in this section will have section_id → NULL (ungrouped).
-        try await section.delete(on: req.db)
-        return req.redirect(to: "/assignments")
-    }
-
-    // MARK: - POST /assignments/:assignmentID/section
-
-    @Sendable
-    func moveToSection(req: Request) async throws -> HTTPStatus {
-        struct MoveBody: Content {
-            var sectionID: String?  // UUID string, or "" / absent = ungrouped
-        }
-        let user = try req.auth.require(APIUser.self)
-        let courseState = try await req.resolveActiveCourse(for: user)
-        guard let courseID = courseState.activeCourseUUID else {
-            throw Abort(.badRequest, reason: "No active course selected.")
-        }
-        let idStr = try assignmentPublicIDParameter(from: req)
-        guard let assignment = try await assignmentByPublicID(idStr, on: req.db),
-              assignment.courseID == courseID else {
-            throw Abort(.notFound)
-        }
-        let body = (try? req.content.decode(MoveBody.self))
-        let newSectionID: UUID? = try await resolveSectionID(body?.sectionID, courseID: courseID, db: req.db)
-        assignment.sectionID = newSectionID
-        try await assignment.save(on: req.db)
-
-        // When moving into a named section, sync the test setup's grading mode
-        // to match the section's defaultGradingMode.  Moving to "ungrouped"
-        // (nil section) leaves the grading mode unchanged.
-        if let sectionUUID = newSectionID,
-           let section     = try await APICourseSection.find(sectionUUID, on: req.db),
-           let setup        = try await APITestSetup.find(assignment.testSetupID, on: req.db) {
-            let mode = section.defaultGradingMode   // "browser" | "worker"
-            if var dict = (try? JSONSerialization.jsonObject(with: Data(setup.manifest.utf8))) as? [String: Any],
-               (dict["gradingMode"] as? String) != mode {
-                dict["gradingMode"] = mode
-                if let data = try? JSONSerialization.data(withJSONObject: dict),
-                   let json = String(data: data, encoding: .utf8) {
-                    setup.manifest = json
-                    try await setup.save(on: req.db)
-                }
-            }
-        }
-
         return .ok
     }
 
@@ -1554,1035 +1035,6 @@ struct AssignmentRoutes: RouteCollection {
             try await assignment.save(on: req.db)
             let q = "notice=\(urlEncode("Assignment updated and validation started. It is still pending; refresh shortly."))"
             return req.redirect(to: "/assignments/\(idStr)/edit?\(q)")
-        }
-    }
-
-    // MARK: - POST /courses/:courseID/open-enrollment
-
-    @Sendable
-    func toggleCourseOpenEnrollment(req: Request) async throws -> Response {
-        struct Body: Content { var openEnrollment: String? }
-        guard
-            let idString = req.parameters.get("courseID"),
-            let courseID = UUID(uuidString: idString),
-            let course   = try await APICourse.find(courseID, on: req.db)
-        else {
-            throw Abort(.notFound)
-        }
-        // An unchecked checkbox sends no field at all (empty body); treat that as false.
-        let body = try? req.content.decode(Body.self)
-        course.openEnrollment = (body?.openEnrollment == "on")
-        try await course.save(on: req.db)
-        return req.redirect(to: "/assignments")
-    }
-
-    // MARK: - POST /courses/:courseID/enroll-csv
-
-    @Sendable
-    func instructorBulkEnrollCSV(req: Request) async throws -> View {
-        struct BulkEnrollForm: Content { var file: Data }
-
-        guard
-            let idString = req.parameters.get("courseID"),
-            let courseID = UUID(uuidString: idString),
-            let course   = try await APICourse.find(courseID, on: req.db),
-            !course.isArchived
-        else {
-            throw Abort(.badRequest, reason: "Invalid or archived course.")
-        }
-
-        let form = try req.content.decode(BulkEnrollForm.self)
-
-        let rawUsernames = parseUsernamesFromCSV(form.file)
-        var seen = Set<String>()
-        let uniqueUsernames = rawUsernames.filter { seen.insert($0).inserted }
-
-        let usernameSet = Set(uniqueUsernames)
-        let allUsers = try await APIUser.query(on: req.db).all()
-        let matchedUsers = allUsers.filter { usernameSet.contains($0.username) }
-
-        let matchedUsernameSet = Set(matchedUsers.map { $0.username })
-        let notFoundUsernames = uniqueUsernames
-            .filter { !matchedUsernameSet.contains($0) }
-            .sorted()
-
-        let existingEnrollments = try await APICourseEnrollment.query(on: req.db)
-            .filter(\.$course.$id == courseID)
-            .all()
-        let alreadyEnrolledUserIDs = Set(existingEnrollments.map { $0.userID })
-
-        var enrolledCount = 0
-        var alreadyEnrolledCount = 0
-
-        for user in matchedUsers {
-            guard let userID = user.id else { continue }
-            if alreadyEnrolledUserIDs.contains(userID) {
-                alreadyEnrolledCount += 1
-            } else {
-                let enrollment = APICourseEnrollment(userID: userID, courseID: courseID)
-                try await enrollment.save(on: req.db)
-                enrolledCount += 1
-            }
-        }
-
-        return try await req.view.render("admin-enroll-csv-result", EnrollCSVResultContext(
-            currentUser:          req.currentUserContext,
-            courseCode:           course.code,
-            courseName:           course.name,
-            enrolledCount:        enrolledCount,
-            alreadyEnrolledCount: alreadyEnrolledCount,
-            notFoundUsernames:    notFoundUsernames,
-            returnURL:            "/assignments"
-        ))
-    }
-}
-
-// MARK: - View context types
-
-struct AssignmentRow: Encodable {
-    let setupID:      String
-    let assignmentID: String?   // nil if unpublished
-    let title:        String?   // nil if unpublished
-    let isOpen:       Bool?     // nil if unpublished
-    let dueAt:        String?
-    let status:       String    // "unpublished" | "open" | "closed"
-    let sortOrder:    Int?
-    let validationStatus: String
-    let validationSubmissionID: String?
-    let suiteCount:   Int
-    let createdAt:    String
-}
-
-/// A course section with its grouped assignment rows, used in instructor and student views.
-struct CourseSectionRow: Encodable {
-    let sectionID: String           // UUID as string
-    let name: String
-    let defaultGradingMode: String  // "browser" | "worker"
-    let sortOrder: Int
-    let rows: [AssignmentRow]       // assignments in this section, sorted
-}
-
-private struct AssignmentsContext: Encodable {
-    let currentUser: CurrentUserContext?
-    let sections: [CourseSectionRow]    // sections with their assignments
-    let ungroupedRows: [AssignmentRow]  // assignments/setups not in any section
-    let hasSections: Bool
-    let hasUngrouped: Bool
-    let enrolledStudents: [EnrolledStudentRow]
-    let hasEnrolledStudents: Bool       // explicit flag — Leaf's array.isEmpty is unreliable
-    let enrolledStudentCount: Int
-    let courseOpenEnrollment: Bool
-    let courseIsArchived: Bool
-}
-
-private struct EnrolledStudentRow: Encodable {
-    let id: String
-    let username: String
-    let displayName: String
-    let role: String        // "student" | "instructor" | "admin"
-}
-
-private struct AssignmentSubmissionsContext: Encodable {
-    let currentUser: CurrentUserContext?
-    let assignmentID: String
-    let assignmentTitle: String
-    let rows: [AssignmentStudentRow]
-}
-
-private struct AssignmentStudentRow: Encodable {
-    let studentID: String
-    let surname: String
-    let givenNames: String
-    let gradeText: String
-    let submissionCount: Int
-    let hasLatestSubmission: Bool
-    let latestSubmissionID: String
-    let latestSubmittedAtText: String
-    let additionalSubmissionCount: Int
-    let fullHistoryURL: String
-}
-
-private struct ValidateContext: Encodable {
-    let currentUser:  CurrentUserContext?
-    let assignmentID: String
-    let setupID:      String
-    let title:        String
-    let suiteCount:   Int
-    let dueAt:        String?
-}
-
-private struct NewAssignmentContext: Encodable {
-    let currentUser: CurrentUserContext?
-    let assignmentName: String
-    let dueAt: String
-    let sections: [CourseSectionRow]    // available sections for the section picker
-    let preselectedSectionID: String    // from ?sectionID= query param
-    let notice: String?
-    let error: String?
-}
-
-private struct EditAssignmentContext: Encodable {
-    let currentUser: CurrentUserContext?
-    let assignmentID: String
-    let assignmentName: String
-    let dueAt: String
-    let currentAssignmentFile: String
-    let currentAssignmentURL: String
-    let currentSolutionFile: String?
-    let currentSolutionURL: String?
-    let existingSuiteRows: [EditableSuiteRow]
-    let notice: String?
-    let error: String?
-}
-
-private struct CurrentFileLink {
-    let name: String
-    let url: String
-}
-
-private struct EditableSuiteRow: Encodable {
-    let name: String
-    let url: String
-    let isTest: Bool
-    let tier: String
-    let order: Int
-}
-
-private struct AssignmentStudentHistoryContext: Encodable {
-    let currentUser: CurrentUserContext?
-    let assignmentID: String
-    let assignmentTitle: String
-    let studentID: String
-    let historyPath: String
-    let rows: [SubmissionHistoryRow]
-}
-
-private struct SubmissionHistoryRow: Encodable {
-    let submissionID: String
-    let attemptNumber: Int
-    let status: String
-    let submittedAt: String
-    let gradeText: String
-}
-
-/// Validates a sectionID string (UUID) against the given course and returns the UUID if valid.
-/// Returns nil for absent, empty, or "none" values (meaning "ungrouped").
-private func resolveSectionID(_ raw: String?, courseID: UUID, db: Database) async throws -> UUID? {
-    guard let raw, !raw.isEmpty, raw.lowercased() != "none" else { return nil }
-    guard let uuid = UUID(uuidString: raw) else {
-        throw Abort(.badRequest, reason: "Invalid sectionID format.")
-    }
-    guard let section = try await APICourseSection.find(uuid, on: db),
-          section.courseID == courseID else {
-        // Section not found or belongs to a different course — silently ignore.
-        return nil
-    }
-    return uuid
-}
-
-private func parseDueDate(_ raw: String?) -> Date? {
-    guard let raw, !raw.isEmpty else { return nil }
-
-    let iso = ISO8601DateFormatter()
-    if let d = iso.date(from: raw) { return d }
-
-    let fmt = DateFormatter()
-    fmt.locale = Locale(identifier: "en_US_POSIX")
-    fmt.dateFormat = "yyyy-MM-dd'T'HH:mm"
-    return fmt.date(from: raw)
-}
-
-private func assignmentByPublicID(_ publicID: String, on db: Database) async throws -> APIAssignment? {
-    try await APIAssignment.query(on: db)
-        .filter(\.$publicID == publicID)
-        .first()
-}
-
-private func isValidAssignmentPublicID(_ value: String) -> Bool {
-    value.count == APIAssignment.publicIDLength
-        && value.allSatisfy { $0.isASCII && ($0.isLetter || $0.isNumber) }
-}
-
-private func assignmentPublicIDParameter(from req: Request) throws -> String {
-    guard let raw = req.parameters.get("assignmentID"), isValidAssignmentPublicID(raw) else {
-        throw Abort(.notFound)
-    }
-    return raw
-}
-
-private func createAssignmentWithUniquePublicID(
-    req: Request,
-    testSetupID: String,
-    title: String,
-    dueAt: Date?,
-    isOpen: Bool,
-    sortOrder: Int?,
-    validationStatus: String? = nil,
-    validationSubmissionID: String? = nil,
-    sectionID: UUID? = nil,
-    courseID: UUID
-) async throws -> APIAssignment {
-    for _ in 0..<32 {
-        let candidate = APIAssignment.generatePublicID()
-        let exists = try await APIAssignment.query(on: req.db)
-            .filter(\.$publicID == candidate)
-            .count() > 0
-        if exists { continue }
-
-        let assignment = APIAssignment(
-            publicID: candidate,
-            testSetupID: testSetupID,
-            title: title,
-            dueAt: dueAt,
-            isOpen: isOpen,
-            sortOrder: sortOrder,
-            validationStatus: validationStatus,
-            validationSubmissionID: validationSubmissionID,
-            sectionID: sectionID,
-            courseID: courseID
-        )
-        do {
-            try await assignment.save(on: req.db)
-        } catch {
-            let conflict = try await APIAssignment.query(on: req.db)
-                .filter(\.$publicID == candidate)
-                .count() > 0
-            if conflict { continue }
-            throw error
-        }
-        return assignment
-    }
-
-    throw Abort(.internalServerError, reason: "Unable to allocate assignment URL id")
-}
-
-private func sanitizedAssignmentReturnPath(
-    _ raw: String?,
-    assignmentIDRaw: String,
-    fallbackPath: String
-) -> String {
-    guard let path = raw?.trimmingCharacters(in: .whitespacesAndNewlines), path.hasPrefix("/") else {
-        return fallbackPath
-    }
-
-    let expectedPrefix = "/assignments/\(assignmentIDRaw)"
-    guard path == expectedPrefix || path.hasPrefix(expectedPrefix + "/") else {
-        return fallbackPath
-    }
-    return path
-}
-
-private func dueAtLocalInputString(_ date: Date?) -> String {
-    guard let date else { return "" }
-    let fmt = DateFormatter()
-    fmt.locale = Locale(identifier: "en_US_POSIX")
-    fmt.dateFormat = "yyyy-MM-dd'T'HH:mm"
-    return fmt.string(from: date)
-}
-
-private func notebookFilenameForStorage(uploadedName: String?, fallback: String) -> String {
-    var fileName = uploadedName?.trimmingCharacters(in: .whitespacesAndNewlines) ?? ""
-    if fileName.isEmpty {
-        fileName = fallback
-    }
-    fileName = URL(fileURLWithPath: fileName).lastPathComponent
-    fileName = fileName
-        .components(separatedBy: CharacterSet(charactersIn: "/\\:*?\"<>|\n\r"))
-        .joined(separator: " ")
-        .trimmingCharacters(in: .whitespacesAndNewlines)
-    if fileName.isEmpty {
-        fileName = fallback
-    }
-    if !fileName.lowercased().hasSuffix(".ipynb") {
-        fileName += ".ipynb"
-    }
-    return fileName
-}
-
-private func currentSetupFiles(for setup: APITestSetup, assignmentID: String, hasValidationSolution: Bool) -> (
-    assignmentFile: CurrentFileLink,
-    solutionFile: CurrentFileLink?,
-    existingSuiteRows: [EditableSuiteRow]
-) {
-    let assignmentFile: CurrentFileLink = {
-        let fileName: String
-        if let path = setup.notebookPath, !path.isEmpty {
-            fileName = URL(fileURLWithPath: path).lastPathComponent
-        } else {
-            fileName = "assignment.ipynb"
-        }
-        return CurrentFileLink(
-            name: fileName,
-            url: "/assignments/\(assignmentID)/files/notebook"
-        )
-    }()
-
-    let manifestSuites: [(script: String, tier: String, order: Int)] = {
-        guard let data = setup.manifest.data(using: .utf8),
-              let props = try? JSONDecoder().decode(TestProperties.self, from: data) else {
-            return []
-        }
-        return props.testSuites.enumerated().map { (idx, item) in
-            (script: item.script, tier: item.tier.rawValue, order: idx + 1)
-        }
-    }()
-    let testMap = Dictionary(uniqueKeysWithValues: manifestSuites.map { ($0.script, $0) })
-
-    let archiveFiles = listZipEntries(zipPath: setup.zipPath)
-    let solutionFile: CurrentFileLink? = {
-        if archiveFiles.contains("solution.ipynb") {
-            return CurrentFileLink(
-                name: "solution.ipynb",
-                url: "/assignments/\(assignmentID)/files/item?name=solution.ipynb"
-            )
-        }
-        if hasValidationSolution {
-            return CurrentFileLink(name: "solution.ipynb", url: "/assignments/\(assignmentID)/files/solution")
-        }
-        return nil
-    }()
-
-    let nonNotebookFiles = archiveFiles
-        .filter { $0 != "assignment.ipynb" && $0 != "solution.ipynb" }
-        .sorted { lhs, rhs in
-            let l = testMap[lhs]?.order ?? Int.max
-            let r = testMap[rhs]?.order ?? Int.max
-            if l != r { return l < r }
-            return lhs < rhs
-        }
-
-    let existingSuiteRows = nonNotebookFiles.enumerated().map { idx, name in
-        let entry = testMap[name]
-        return EditableSuiteRow(
-            name: name,
-            url: "/assignments/\(assignmentID)/files/item?name=\(urlEncode(name))",
-            isTest: entry != nil,
-            tier: entry?.tier ?? "support",
-            order: entry?.order ?? (idx + 1)
-        )
-    }
-
-    return (assignmentFile, solutionFile, existingSuiteRows)
-}
-
-private func listZipEntries(zipPath: String) -> [String] {
-    let process = Process()
-    process.executableURL = URL(fileURLWithPath: "/usr/bin/unzip")
-    process.arguments = ["-Z1", zipPath]
-
-    let out = Pipe()
-    process.standardOutput = out
-    process.standardError = Pipe()
-
-    do {
-        try process.run()
-    } catch {
-        return []
-    }
-    let data = out.fileHandleForReading.readDataToEndOfFile()
-    process.waitUntilExit()
-    guard process.terminationStatus == 0 else { return [] }
-    guard let text = String(data: data, encoding: .utf8) else { return [] }
-    return text
-        .split(separator: "\n")
-        .map(String.init)
-        .map { $0.trimmingCharacters(in: .whitespacesAndNewlines) }
-        .filter { !$0.isEmpty && !$0.hasSuffix("/") }
-}
-
-private func extractZipEntry(zipPath: String, entryName: String) -> Data? {
-    let process = Process()
-    process.executableURL = URL(fileURLWithPath: "/usr/bin/unzip")
-    process.arguments = ["-p", zipPath, entryName]
-    let out = Pipe()
-    process.standardOutput = out
-    process.standardError = Pipe()
-    do {
-        try process.run()
-    } catch {
-        return nil
-    }
-    let data = out.fileHandleForReading.readDataToEndOfFile()
-    process.waitUntilExit()
-    guard process.terminationStatus == 0 else { return nil }
-    return data
-}
-
-private func buildFileResponse(data: Data, filename: String) -> Response {
-    var headers = HTTPHeaders()
-    headers.contentType = contentType(for: filename)
-    headers.add(name: .contentDisposition, value: "attachment; filename=\"\(filename)\"")
-    return Response(status: .ok, headers: headers, body: .init(data: data))
-}
-
-private func contentType(for filename: String) -> HTTPMediaType {
-    switch URL(fileURLWithPath: filename).pathExtension.lowercased() {
-    case "ipynb", "json":
-        return .json
-    case "py", "r", "sh", "bash", "zsh", "rb", "pl", "js", "php", "txt", "md", "csv":
-        return .plainText
-    default:
-        return HTTPMediaType(type: "application", subType: "octet-stream")
-    }
-}
-
-private struct EditSuiteConfigRow: Decodable {
-    let source: String?
-    let name: String?
-    let index: Int?
-    let isIncluded: Bool?
-    let isTest: Bool?
-    let tier: String?
-    let order: Int?
-}
-
-private struct ReindexedSuiteConfigRow: Encodable {
-    let index: Int
-    let isTest: Bool
-    let tier: String
-    let order: Int?
-}
-
-private struct ResolvedEditSuiteFiles {
-    let files: [File]
-    let reindexedSuiteConfigJSON: String?
-}
-
-private func resolveEditSuiteFiles(
-    setupZipPath: String,
-    setupManifestJSON: String,
-    uploadedSuiteFiles: [File],
-    suiteConfigJSON: String?
-) throws -> ResolvedEditSuiteFiles {
-    let parsedRows: [EditSuiteConfigRow] = {
-        guard let raw = suiteConfigJSON?.trimmingCharacters(in: .whitespacesAndNewlines),
-              !raw.isEmpty,
-              let data = raw.data(using: .utf8),
-              let rows = try? JSONDecoder().decode([EditSuiteConfigRow].self, from: data) else {
-            return []
-        }
-        return rows
-    }()
-
-    // Backward compatibility: no table config submitted.
-    // Preserve existing suite/support files and append any new uploads.
-    if parsedRows.isEmpty {
-        let existingEntries = listZipEntries(zipPath: setupZipPath)
-            .filter { $0 != "assignment.ipynb" && $0 != "solution.ipynb" }
-            .sorted()
-
-        var resolvedFiles: [File] = []
-        var configRows: [ReindexedSuiteConfigRow] = []
-        var nextOrder = 1
-
-        let manifestTests: [String: (tier: String, order: Int)] = {
-            guard let data = setupManifestJSON.data(using: .utf8),
-                  let props = try? JSONDecoder().decode(TestProperties.self, from: data) else {
-                return [:]
-            }
-            var map: [String: (tier: String, order: Int)] = [:]
-            for (idx, entry) in props.testSuites.enumerated() {
-                map[entry.script] = (entry.tier.rawValue, idx + 1)
-            }
-            return map
-        }()
-
-        for name in existingEntries {
-            guard let data = extractZipEntry(zipPath: setupZipPath, entryName: name) else { continue }
-            var buffer = ByteBufferAllocator().buffer(capacity: data.count)
-            buffer.writeBytes(data)
-            resolvedFiles.append(File(data: buffer, filename: name))
-
-            let testInfo = manifestTests[name]
-            let tier = testInfo?.tier ?? "support"
-            configRows.append(ReindexedSuiteConfigRow(
-                index: resolvedFiles.count - 1,
-                isTest: testInfo != nil && tier != "support",
-                tier: tier,
-                order: testInfo?.order ?? nextOrder
-            ))
-            nextOrder += 1
-        }
-
-        let appendedUploads = uploadedSuiteFiles.filter { $0.data.readableBytes > 0 }
-        for (idx, file) in appendedUploads.enumerated() {
-            let rawName = file.filename.isEmpty ? "suite-file-\(idx + 1)" : file.filename
-            let cleanName = sanitizeSuiteFilename(rawName)
-            let data = Data(file.data.readableBytesView)
-            guard !data.isEmpty else { continue }
-            var buffer = ByteBufferAllocator().buffer(capacity: data.count)
-            buffer.writeBytes(data)
-            resolvedFiles.append(File(data: buffer, filename: cleanName))
-
-            let ext = URL(fileURLWithPath: cleanName).pathExtension.lowercased()
-            let likelyTest = ["sh","bash","zsh","py","rb","pl","js","php"].contains(ext)
-            configRows.append(ReindexedSuiteConfigRow(
-                index: resolvedFiles.count - 1,
-                isTest: likelyTest,
-                tier: likelyTest ? "public" : "support",
-                order: nextOrder
-            ))
-            nextOrder += 1
-        }
-
-        let configJSON: String? = {
-            guard let data = try? JSONEncoder().encode(configRows) else { return nil }
-            return String(data: data, encoding: .utf8)
-        }()
-        return ResolvedEditSuiteFiles(
-            files: resolvedFiles,
-            reindexedSuiteConfigJSON: configJSON
-        )
-    }
-
-    var resolvedFiles: [File] = []
-    var configRows: [ReindexedSuiteConfigRow] = []
-    var nextOrder = 1
-
-    for row in parsedRows {
-        let included = row.isIncluded ?? true
-        guard included else { continue }
-
-        let source = (row.source ?? "").lowercased()
-        let dataAndName: (Data, String)?
-        if source == "existing" {
-            guard let rawName = row.name, !rawName.isEmpty else { continue }
-            let cleanName = (rawName as NSString).lastPathComponent
-            guard cleanName == rawName, !cleanName.isEmpty else { continue }
-            guard let data = extractZipEntry(zipPath: setupZipPath, entryName: cleanName) else { continue }
-            dataAndName = (data, cleanName)
-        } else if source == "upload" {
-            guard let idx = row.index, uploadedSuiteFiles.indices.contains(idx) else { continue }
-            let file = uploadedSuiteFiles[idx]
-            let data = Data(file.data.readableBytesView)
-            guard !data.isEmpty else { continue }
-            let rawName = file.filename.isEmpty ? "suite-file-\(idx + 1)" : file.filename
-            dataAndName = (data, sanitizeSuiteFilename(rawName))
-        } else {
-            continue
-        }
-
-        guard let (data, name) = dataAndName else { continue }
-
-        var buffer = ByteBufferAllocator().buffer(capacity: data.count)
-        buffer.writeBytes(data)
-        resolvedFiles.append(File(data: buffer, filename: name))
-
-        let tier = normalizeTier(row.tier)
-        let isTest = (row.isTest ?? false) && tier != "support"
-        configRows.append(ReindexedSuiteConfigRow(
-            index: resolvedFiles.count - 1,
-            isTest: isTest,
-            tier: tier,
-            order: row.order ?? nextOrder
-        ))
-        nextOrder += 1
-    }
-
-    let configJSON: String? = {
-        guard let data = try? JSONEncoder().encode(configRows) else { return nil }
-        return String(data: data, encoding: .utf8)
-    }()
-    return ResolvedEditSuiteFiles(files: resolvedFiles, reindexedSuiteConfigJSON: configJSON)
-}
-
-private func loadExistingSolutionNotebook(req: Request, assignment: APIAssignment) async throws -> Data? {
-    if let validationID = assignment.validationSubmissionID,
-       let validationSubmission = try await APISubmission.find(validationID, on: req.db),
-       let data = try? Data(contentsOf: URL(fileURLWithPath: validationSubmission.zipPath)),
-       !data.isEmpty {
-        return data
-    }
-
-    if let fallbackSubmission = try await APISubmission.query(on: req.db)
-        .filter(\.$testSetupID == assignment.testSetupID)
-        .filter(\.$kind == APISubmission.Kind.validation)
-        .filter(\.$filename == "solution.ipynb")
-        .sort(\.$submittedAt, .descending)
-        .first(),
-       let data = try? Data(contentsOf: URL(fileURLWithPath: fallbackSubmission.zipPath)),
-       !data.isEmpty {
-        return data
-    }
-
-    return nil
-}
-
-private func urlEncode(_ s: String) -> String {
-    var allowed = CharacterSet.alphanumerics
-    allowed.insert(charactersIn: "-._~")
-    return s.addingPercentEncoding(withAllowedCharacters: allowed) ?? ""
-}
-
-private func nextAssignmentSortOrder(req: Request) async throws -> Int {
-    let maxOrder = try await APIAssignment.query(on: req.db)
-        .all()
-        .compactMap(\.sortOrder)
-        .max() ?? 0
-    return maxOrder + 1
-}
-
-private func gradePercentFromCollectionJSON(_ collectionJSON: String) -> Int? {
-    guard let data = collectionJSON.data(using: .utf8),
-          let root = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-          let passCount = root["passCount"] as? Int,
-          let totalTests = root["totalTests"] as? Int,
-          totalTests > 0 else {
-        return nil
-    }
-    return Int((Double(passCount) / Double(totalTests) * 100).rounded())
-}
-
-private func csvEscaped(_ value: String) -> String {
-    if value.contains(",") || value.contains("\"") || value.contains("\n") || value.contains("\r") {
-        return "\"" + value.replacingOccurrences(of: "\"", with: "\"\"") + "\""
-    }
-    return value
-}
-
-private func inferNameFromStudentID(_ studentID: String) -> (surname: String, givenNames: String) {
-    let raw = studentID.trimmingCharacters(in: .whitespacesAndNewlines)
-    guard !raw.isEmpty else { return ("—", "—") }
-
-    if raw.contains(",") {
-        let parts = raw.split(separator: ",", maxSplits: 1, omittingEmptySubsequences: false)
-        let surname = String(parts.first ?? "").trimmingCharacters(in: .whitespacesAndNewlines)
-        let given = parts.count > 1
-            ? String(parts[1]).trimmingCharacters(in: .whitespacesAndNewlines)
-            : ""
-        return (
-            surname.isEmpty ? "—" : surname,
-            given.isEmpty ? "—" : given
-        )
-    }
-    return ("—", "—")
-}
-
-private func defaultNotebookData(title: String) -> Data {
-    let safeTitle = title.replacingOccurrences(of: "\"", with: "\\\"")
-    let json = """
-    {
-      "cells": [
-        {
-          "cell_type": "markdown",
-          "metadata": {},
-          "source": ["# \(safeTitle)\\n", "\\n", "Write your assignment instructions here.\\n"]
-        },
-        {
-          "cell_type": "code",
-          "execution_count": null,
-          "metadata": {},
-          "outputs": [],
-          "source": ["# Student solution starts here\\n"]
-        }
-      ],
-      "metadata": {
-        "kernelspec": {
-          "display_name": "Python (Pyodide)",
-          "language": "python",
-          "name": "python"
-        },
-        "language_info": {
-          "name": "python"
-        }
-      },
-      "nbformat": 4,
-      "nbformat_minor": 5
-    }
-    """
-    return Data(json.utf8)
-}
-
-private func createRunnerSetupZip(
-    assignmentNotebookData: Data,
-    solutionNotebookData: Data?,
-    suiteFiles: [File],
-    suiteConfigJSON: String?,
-    zipPath: String
-) throws -> RunnerSetupPackage {
-    let fm = FileManager.default
-    let tempDir = fm.temporaryDirectory.appendingPathComponent("chickadee_runner_setup_\(UUID().uuidString)")
-    try fm.createDirectory(at: tempDir, withIntermediateDirectories: true)
-    defer { try? fm.removeItem(at: tempDir) }
-
-    var seenNames: Set<String> = []
-    var storedNameByIndex: [Int: String] = [:]
-    for (index, file) in suiteFiles.enumerated() {
-        let data = Data(file.data.readableBytesView)
-        guard !data.isEmpty else { continue }
-        let rawName = file.filename.isEmpty ? "suite-file-\(index + 1)" : file.filename
-        let baseName = sanitizeSuiteFilename(rawName)
-        let finalName: String
-        if !seenNames.contains(baseName) {
-            finalName = baseName
-        } else {
-            let ext = URL(fileURLWithPath: baseName).pathExtension
-            let stem = (baseName as NSString).deletingPathExtension
-            var suffix = 2
-            var candidate = baseName
-            while seenNames.contains(candidate) {
-                candidate = ext.isEmpty ? "\(stem)-\(suffix)" : "\(stem)-\(suffix).\(ext)"
-                suffix += 1
-            }
-            finalName = candidate
-        }
-        seenNames.insert(finalName)
-        try data.write(to: tempDir.appendingPathComponent(finalName))
-        storedNameByIndex[index] = finalName
-    }
-
-    let notebookURL = tempDir.appendingPathComponent("assignment.ipynb")
-    try assignmentNotebookData.write(to: notebookURL)
-    if let solutionNotebookData {
-        let solutionURL = tempDir.appendingPathComponent("solution.ipynb")
-        try solutionNotebookData.write(to: solutionURL)
-    }
-
-    let testSuites = try buildSuiteEntries(
-        suiteFiles: suiteFiles,
-        storedNameByIndex: storedNameByIndex,
-        suiteConfigJSON: suiteConfigJSON
-    )
-    guard !testSuites.isEmpty else {
-        throw Abort(.badRequest, reason: "Select at least one test file in the suite file list")
-    }
-
-    let zip = Process()
-    zip.executableURL = URL(fileURLWithPath: "/usr/bin/zip")
-    zip.currentDirectoryURL = tempDir
-    zip.arguments = ["-q", "-r", zipPath, "."]
-    try zip.run()
-    zip.waitUntilExit()
-    guard zip.terminationStatus == 0 else {
-        throw Abort(.internalServerError, reason: "Failed to package setup zip")
-    }
-    let hasMakefile = storedNameByIndex.values.contains {
-        let n = $0.lowercased()
-        return n == "makefile" || n == "gnumakefile"
-    }
-    return RunnerSetupPackage(testSuites: testSuites, hasMakefile: hasMakefile)
-}
-
-private func sanitizeSuiteFilename(_ raw: String) -> String {
-    var name = (raw as NSString).lastPathComponent
-    if name.isEmpty { name = "suite-file" }
-    name = name.replacingOccurrences(of: "/", with: "-")
-    name = name.replacingOccurrences(of: "\\", with: "-")
-    return name
-}
-
-private struct SuiteConfigRow: Decodable {
-    let index: Int
-    let isTest: Bool
-    let tier: String?
-    let order: Int?
-}
-
-private struct ConfiguredSuiteEntry {
-    let script: String
-    let tier: String
-    let order: Int
-}
-
-private struct RunnerSetupPackage {
-    let testSuites: [ConfiguredSuiteEntry]
-    let hasMakefile: Bool
-}
-
-private func buildSuiteEntries(
-    suiteFiles: [File],
-    storedNameByIndex: [Int: String],
-    suiteConfigJSON: String?
-) throws -> [ConfiguredSuiteEntry] {
-    let parsedRows: [SuiteConfigRow] = {
-        guard let raw = suiteConfigJSON?.trimmingCharacters(in: .whitespacesAndNewlines),
-              !raw.isEmpty,
-              let data = raw.data(using: .utf8),
-              let rows = try? JSONDecoder().decode([SuiteConfigRow].self, from: data) else {
-            return []
-        }
-        return rows
-    }()
-
-    if !parsedRows.isEmpty {
-        var rowsByIndex: [Int: SuiteConfigRow] = [:]
-        for row in parsedRows {
-            rowsByIndex[row.index] = row
-        }
-        var selected: [ConfiguredSuiteEntry] = []
-        for index in suiteFiles.indices {
-            guard let row = rowsByIndex[index], row.isTest else { continue }
-            guard let script = storedNameByIndex[index], !script.isEmpty else { continue }
-            let tier = normalizeTier(row.tier)
-            selected.append(ConfiguredSuiteEntry(
-                script: script,
-                tier: tier,
-                order: row.order ?? (index + 1)
-            ))
-        }
-        return selected
-            .sorted { lhs, rhs in
-                if lhs.order != rhs.order { return lhs.order < rhs.order }
-                return lhs.script < rhs.script
-            }
-    }
-
-    // Backward-compatible fallback when no suite config JSON is submitted.
-    let supportedExtensions: Set<String> = ["sh", "bash", "zsh", "py", "r", "rb", "pl", "js", "php"]
-    var defaults: [ConfiguredSuiteEntry] = []
-    for index in suiteFiles.indices {
-        guard let script = storedNameByIndex[index], !script.isEmpty else { continue }
-        let ext = URL(fileURLWithPath: script).pathExtension.lowercased()
-        guard supportedExtensions.contains(ext) else { continue }
-        defaults.append(ConfiguredSuiteEntry(
-            script: script,
-            tier: "public",
-            order: inferredOrder(from: script) ?? (index + 1)
-        ))
-    }
-    return defaults
-        .sorted { lhs, rhs in
-            if lhs.order != rhs.order { return lhs.order < rhs.order }
-            return lhs.script < rhs.script
-        }
-}
-
-private func inferredOrder(from filename: String) -> Int? {
-    let base = (filename as NSString).lastPathComponent
-    let ns = base as NSString
-    let range = NSRange(location: 0, length: ns.length)
-    let regex = try? NSRegularExpression(pattern: #"^([0-9]+)[_-].+$"#)
-    guard let match = regex?.firstMatch(in: base, options: [], range: range),
-          match.numberOfRanges >= 2,
-          let orderRange = Range(match.range(at: 1), in: base) else {
-        return nil
-    }
-    return Int(base[orderRange])
-}
-
-private func normalizeTier(_ raw: String?) -> String {
-    switch (raw ?? "public").lowercased() {
-    case "secret": return "secret"
-    case "release": return "release"
-    case "student": return "student"
-    default: return "public"
-    }
-}
-
-private func makeWorkerManifestJSON(
-    testSuites: [ConfiguredSuiteEntry],
-    includeMakefile: Bool,
-    gradingMode: String = "worker"
-) throws -> String {
-    let testSuiteJSON = testSuites.map { ["tier": $0.tier, "script": $0.script] }
-    let manifest: [String: Any] = [
-        "schemaVersion": 1,
-        "gradingMode": gradingMode,
-        "requiredFiles": [],
-        "testSuites": testSuiteJSON,
-        "timeLimitSeconds": 10,
-        "makefile": includeMakefile ? ["target": NSNull()] : NSNull()
-    ]
-    let data = try JSONSerialization.data(withJSONObject: manifest)
-    return String(data: data, encoding: .utf8) ?? "{}"
-}
-
-private func enqueueRunnerValidationSubmission(
-    req: Request,
-    setupID: String,
-    solutionNotebookData: Data
-) async throws -> String {
-    let submissionsDir = req.application.submissionsDirectory
-    let subID = "sub_\(UUID().uuidString.lowercased().prefix(8))"
-    let filePath = submissionsDir + "\(subID).ipynb"
-    try solutionNotebookData.write(to: URL(fileURLWithPath: filePath))
-
-    let priorCount = try await APISubmission.query(on: req.db)
-        .filter(\.$testSetupID == setupID)
-        .filter(\.$kind == APISubmission.Kind.validation)
-        .count()
-
-    let user = try req.auth.require(APIUser.self)
-    let submission = APISubmission(
-        id:            subID,
-        testSetupID:   setupID,
-        zipPath:       filePath,
-        attemptNumber: priorCount + 1,
-        filename:      "solution.ipynb",
-        userID:        user.id,
-        kind:          APISubmission.Kind.validation
-    )
-    try await submission.save(on: req.db)
-    return subID
-}
-
-private enum RunnerValidationOutcome {
-    case passed(summary: String)
-    case failed(summary: String)
-    case timedOut
-}
-
-private func waitForRunnerValidation(
-    req: Request,
-    submissionID: String,
-    timeoutSeconds: TimeInterval = 45
-) async throws -> RunnerValidationOutcome {
-    let started = Date()
-    let decoder = JSONDecoder()
-    decoder.dateDecodingStrategy = .iso8601
-
-    while Date().timeIntervalSince(started) < timeoutSeconds {
-        guard let submission = try await APISubmission.find(submissionID, on: req.db),
-              submission.kind == APISubmission.Kind.validation else {
-            throw Abort(.notFound, reason: "Validation submission missing")
-        }
-
-        if submission.status == "complete" || submission.status == "failed" {
-            guard let result = try await APIResult.query(on: req.db)
-                .filter(\.$submissionID == submissionID)
-                .sort(\.$receivedAt, .descending)
-                .first(),
-                  let data = result.collectionJSON.data(using: .utf8) else {
-                return .failed(summary: "no result payload")
-            }
-
-            let collection = try decoder.decode(TestOutcomeCollection.self, from: data)
-            let summary = "\(collection.passCount)/\(collection.totalTests) passed"
-            let passed = collection.buildStatus == .passed &&
-                collection.failCount == 0 &&
-                collection.errorCount == 0 &&
-                collection.timeoutCount == 0
-            return passed ? .passed(summary: summary) : .failed(summary: summary)
-        }
-
-        try await Task.sleep(nanoseconds: 1_000_000_000)
-    }
-
-    return .timedOut
-}
-
-private func ensureValidationRunnerAvailability(req: Request) async {
-    let enabled = await req.application.localRunnerAutoStartStore.isEnabled()
-    guard enabled else { return }
-
-    let hasRecentRunner = await req.application.workerActivityStore.hasRecentActivity(within: 20)
-    guard !hasRecentRunner else { return }
-
-    await req.application.localRunnerManager.ensureRunning(app: req.application, logger: req.logger)
-    try? await Task.sleep(nanoseconds: 1_000_000_000)
-}
-
-private func removeMaterializedNotebookFiles(req: Request, setupID: String) {
-    let roots = [
-        req.application.directory.publicDirectory + "files/",
-        req.application.directory.publicDirectory + "jupyterlite/files/",
-        req.application.directory.publicDirectory + "jupyterlite/lab/files/",
-        req.application.directory.publicDirectory + "jupyterlite/notebooks/files/"
-    ]
-    for root in roots {
-        guard let entries = try? FileManager.default.contentsOfDirectory(atPath: root) else { continue }
-        for name in entries where name.hasPrefix(setupID) && name.hasSuffix(".ipynb") {
-            try? FileManager.default.removeItem(atPath: root + name)
         }
     }
 }
